Static Configuration Files
These files are meant to be copied directly into place and then used by
scripts to be committed later. May include a setup script to handle the
copy step in a later commit.

For the moment, the `nginx/` directory should be merged into your
existing Nginx configuration directory. The `ssl/` directory should
similarly be merged into your SSL configuration directory.

If you're using Windows, also change the `/etc/ssl/` portions in
`nginx/nanobox.app` to match the location of your SSL configuration
directory.

The commands which make up the majority of the upcoming scripts can be
found at https://content.nanobox.io/testing-https-locally-with-nanobox/
(as soon as that article is published).

This commit message will probably be adapted into a README in a future
commit.